refactor(layout): Use `src` as the base directory
This should clean up the `../` hell.
