+++ src/components/controls/Footer.astro
@@ -1,5 +1,5 @@
 ---
-import FooterIcon from "./FooterIcon.astro";
+import FooterIcon from "components/controls/FooterIcon.astro";
 ---
 
 <footer>
