+++ src/components/notes/NoteHeader.astro
@@ -1,6 +1,6 @@
 ---
-import type { Tag as TagType } from "../../data/tags";
-import Tag from "../tags/Tag.astro";
+import type { Tag as TagType } from "stores/tags";
+import Tag from "components/tags/Tag.astro";
 
 type Props = {
 	tagDescriptions: TagType[];
