+++ src/components/notes/NoteList.astro
@@ -1,6 +1,6 @@
 ---
-import { classes } from "../../utils/classes";
-import type { Note } from "../../data/notes";
+import type { Note } from "stores/notes";
+import { classes } from "utils/classes";
 import { DateTime } from "luxon";
 
 type Props = {
