+++ src/components/tags/Tag.astro
@@ -1,9 +1,9 @@
 ---
-import { classes } from "../../utils/classes";
+import { classes } from "utils/classes";
 
 type Props = {
 	description?: string;
-	large: boolean;
+	large?: boolean;
 	href?: string;
 };
