+++ src/components/tags/TagHeader.astro
@@ -1,5 +1,5 @@
 ---
-import Tag from "./Tag.astro";
+import Tag from "components/tags/Tag.astro";
 
 type Props = {
 	description?: string;
