+++ src/components/tags/TagListing.astro
@@ -1,9 +1,9 @@
 ---
-import ButtonLink from "../controls/ButtonLink.astro";
-import NoteList from "../notes/NoteList.astro";
-import type { Note } from "../../data/notes";
+import ButtonLink from "components/controls/ButtonLink.astro";
+import NoteList from "components/notes/NoteList.astro";
+import Row from "components/controls/Row.astro";
 import TagHeader from "./TagHeader.astro";
-import Row from "../controls/Row.astro";
+import type { Note } from "stores/notes";
 
 type Props = {
 	description?: string;
