+++ src/pages/404.mdx
@@ -1,5 +1,5 @@
-import Hero from "../components/controls/Hero.astro";
-import Root from "../layouts/Root.astro";
+import Hero from "components/controls/Hero.astro";
+import Root from "layouts/Root.astro";
 
 <Root title="Page not found">
     <Hero title="You look lost.">
