+++ src/pages/projects.mdx
@@ -1,5 +1,5 @@
-import Hero from "../components/controls/Hero.astro";
-import Root from "../layouts/Root.astro";
+import Hero from "components/controls/Hero.astro";
+import Root from "layouts/Root.astro";
 
 <Root
     description={
