+++ src/stores/notes.ts
