+++ tsconfig.json
@@ -1,6 +1,9 @@
 {
-  "extends": "astro/tsconfigs/strict",
-  "include": [
-    "src/**/*"
-  ]
+	"compilerOptions": {
+		"baseUrl": "src"
+	},
+	"extends": "astro/tsconfigs/strict",
+	"include": [
+		"src/**/*"
+	]
 }
